% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weights.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Function for calculating weights for logrank test}
\usage{
calculate_weights(risk_table, method = "logrank", hr_fun = NULL, rho = 0,
  gamma = 0)
}
\arguments{
\item{risk_table}{Risk table from the function \code{}}

\item{method}{Which method for calculating the weights
\code{logrank} Standard logrank test (w=1 for all)
\code{theta} w = -log(hr(t))
\code{hr_weight} Weight from article
\code{fh} Flemming-Harrington weights}

\item{hr_fun}{Hazard ratio function for method \code{theta} and \code{hr_weight}}

\item{rho}{Variable for method \code{fh}}

\item{gamma}{Varaiable for method \code{fh}}
}
\value{
Risk table updated with weights
}
\description{
Calculates weights for treatment switching logrank test
}
