% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_zs.R
\name{calculate_zs}
\alias{calculate_zs}
\title{Get (standardized) score and weighted log-rank statistics.}
\usage{
calculate_zs(risk_table)
}
\arguments{
\item{risk_table}{A risk table with weights produced by the functions \code{get_risk_table} and \code{calculate_weights}.}
}
\value{
The standardized weighted log-rank statistic.
}
\description{
\code{get_zs} Calculate (standardized) score and weighted log-rank statistics from risk table (with weights).
}
