% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_fun.R
\name{exp_prog_switch_fun}
\alias{exp_prog_switch_fun}
\title{Creates survival and hazard ratio function for treatment switching with exponential model}
\usage{
exp_prog_switch_fun(lambda_c, lambda_e, lambda_p, p)
}
\arguments{
\item{lambda_c}{rate for control group}

\item{lambda_e}{rate for experimental group}

\item{lambda_p}{rate for progression in control group}

\item{p}{proportion of patients that switches after progression}
}
\value{
a list containing functions for the survival curve and HR and hazard for control and proportion of switchers
}
\description{
\code{exp_prog_switch_fun} numerically estimates the survival and HR functions in a treatment
switching model with exponentially distributions
}
