% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_fun.R
\name{HR_delay_switch}
\alias{HR_delay_switch}
\title{Calculates the hazard ratio for switching treatment at a delay time}
\usage{
HR_delay_switch(lambda_c_1, lambda_c_2, lambda_e, delay, p)
}
\arguments{
\item{lambda_c_1}{Hazard rate in control group before switching}

\item{lambda_c_2}{Hazard rate in control group after switching}

\item{lambda_e}{Hazard rate in experimental group}

\item{delay}{Time where treatment changes}

\item{p}{Proportion of (alive) patients that switches treatment at time delay}
}
\value{
Survival and hazard ratio function
}
\description{
\code{HR_delay_switch} Hazard ratio function for switching treatment at a delay time
}
