% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_delay_crossover.R
\name{MC_delay_crossover}
\alias{MC_delay_crossover}
\title{Monte Carlo simulation for treatment switching with delay time switching}
\usage{
MC_delay_crossover(n_c = 100, n_e = 100, median_c_1 = 10,
  median_c_2 = 10, median_e = 15, delay = 5, end_event = 150,
  rec_period = 12, rec_power = 1, p = c(0, 1), alpha = 0.025, M = 100)
}
\arguments{
\item{n_c}{Number of patients in control}

\item{n_e}{Number of patients in experimental}

\item{median_c_1}{Median in control before switching (and for patients that do not switch)}

\item{median_c_2}{Median in control after switching (for the patients that switches)}

\item{delay}{Delay time for when switching occurs}

\item{end_event}{Number of events in the study}

\item{rec_period}{Time for recruitment}

\item{rec_power}{Variable for recruitment that follows the power model}

\item{p}{Proportion(s) of patients that switches after the delay}

\item{alpha}{One-sided significance level}

\item{M}{Number of simulations (for each p)}

\item{median_e_1}{Median in experimental}
}
\description{
Runs a Monte Carlo simulation for the treatment switching model with switch after delay
}
