% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_exp_prog_crossover.R
\name{MC_exp_prog_crossover}
\alias{MC_exp_prog_crossover}
\title{Monte Carlo simulation for treatment switching with switching after progression}
\usage{
MC_exp_prog_crossover(n_c = 100, n_e = 100, median_c = 10,
  median_e = 15, median_prog = 5, end_event = 150, rec_period = 12,
  rec_power = 1, p = c(0, 1), alpha = 0.025, M = 100)
}
\arguments{
\item{n_c}{Number of patients in control}

\item{n_e}{Number of patients in experimental}

\item{median_c}{Median in control arm}

\item{median_e}{Median in experimental arm}

\item{median_prog}{Median for progressions}

\item{end_event}{Number of events to stop the study}

\item{rec_period}{Time for recruitment}

\item{rec_power}{Variable for recruitment that follows the power model}

\item{p}{Proportion(s) of patients that switches after the delay}

\item{alpha}{One-sided significance level}

\item{M}{Number of simulations (for each p)}
}
\description{
Runs a Monte Carlo simulation with treatment switching after progression
}
