% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_fun.R
\name{numerical_prog_switch_exp}
\alias{numerical_prog_switch_exp}
\title{Creates numerically estimated survival and hazard ratio function for treatment switching with exponential model}
\usage{
numerical_prog_switch_exp(lambda_c, lambda_e, lambda_p, p, max_t = 100,
  delta_t = 0.01)
}
\arguments{
\item{lambda_c}{rate for control group}

\item{lambda_e}{rate for experimental group}

\item{lambda_p}{rate for progression in control group}

\item{p}{proportion of patients that switches after progression}

\item{max_t}{maximum time in time interval}

\item{delta_t}{size of time intervals in numerical approximations}
}
\value{
a list containing functions for the survival curve and HR and hazard for control and proportion of switchers
}
\description{
\code{numerical_prog_switch_exp} numerically estimates the survival and HR functions in a treatment
switching model with exponentially distributions
}
