% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_switch.R
\name{prop_switch}
\alias{prop_switch}
\title{Calculates the propoertion of patients that have switched treatment in 1 simulation}
\usage{
prop_switch(dt)
}
\arguments{
\item{dt}{A simulation with treatment switching}
}
\value{
Proportion of patients that have switched treatment
}
\description{
Calculates the propoertion of patients that have switched treatment in 1 simulation
}
