% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_crossover_delay.R
\name{sim_crossover_delay}
\alias{sim_crossover_delay}
\title{Simulation with treatment switching at a time}
\usage{
sim_crossover_delay(n_c = 100, n_e = 100, median_c_1 = 10,
  median_c_2 = 12.5, median_e = 15, delay = 3, p_change = 1,
  end_event = 150, rec_period = 12, rec_power = 1, model = NULL)
}
\arguments{
\item{n_c}{Number of patients in control arm}

\item{n_e}{Number of patients in experimental arm}

\item{median_c_1}{Median in control arm for the first time period}

\item{median_c_2}{Median in control arm for the second time period}

\item{median_e}{Median in experimental arm}

\item{delay}{Time at which to change treatment in control}

\item{p_change}{Probability that a patient in control that has a progression changes to experimental arm}

\item{end_event}{Number of events to stop the study}

\item{rec_period}{Recruitment period}

\item{rec_power}{Recruitment follows a power model, this variable how the power looks}

\item{model}{List containing all the variabels above}
}
\value{
dt Data table containing time, event and group/arm
}
\description{
\code{sim_crossover_delay} Simulating a trial with crossover effect.
Control and experimental arm are both exponentially distributed,
but control switches treatment after a delay time
}
