% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exp_crossover.R
\name{sim_exp_crossover}
\alias{sim_exp_crossover}
\title{Simulation with crossover effect}
\usage{
sim_exp_crossover(n_c = 100, n_e = 100, median_c = 10, median_e = 15,
  median_progression = 5, p_change = 0.5, end_event = 150,
  rec_period = 12, rec_power = 1, model = NULL)
}
\arguments{
\item{n_c}{Number of patients in control arm}

\item{n_e}{Number of patients in experimental arm}

\item{median_c}{Median in control arm}

\item{median_e}{Median in experimental arm}

\item{median_progression}{Median for getting a progression in control arm}

\item{p_change}{Probability that a patient in control that has a progression changes to experimental arm}

\item{end_event}{Number of events to stop the study}

\item{rec_period}{Recruitment period}

\item{rec_power}{Recruitment follows a power model, this variable how the power looks}

\item{model}{List containing all the variabels above}
}
\value{
dt Data table containing time, event and group/arm
}
\description{
\code{sim_exp_crossover} Simulating a trial with crossover effect.
Control and experimental arm are both exponentially distributed,
but control can have progression and might change to experimental arm
}
